verbs: Move atomic functions into separate file
Isolate atomic interfaces implementation into their own
header file.
